requirements.txt
numpy
pandas
screeninfo
pyyaml
playsound

psychopy